Removed HasUrl interface and trait
