+++ src/Activatable.php
@@ -17,6 +17,23 @@ interface Activatable
     public function setInactive();
 
     /**
+     * @return string|null
+     */
+    public function url();
+
+    /**
+     * @return bool
+     */
+    public function hasUrl(): bool;
+
+    /**
+     * @param string|null $url
+     *
+     * @return $this
+     */
+    public function setUrl($url);
+
+    /**
      * @param string $url
      * @param string $root
      */
