--- src/HasUrl.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Spatie\Menu;
-
-interface HasUrl extends Activatable
-{
-    /**
-     * @return string
-     */
-    public function url(): string;
-
-    /**
-     * @param string $url
-     *
-     * @return $this
-     */
-    public function setUrl(string $url);
-}
